+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/ExceptionDialogBuilder.java
@@ -24,7 +24,6 @@ import is.codion.common.property.PropertyValue;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.UndeclaredThrowableException;
-import java.rmi.RemoteException;
 import java.util.Collection;
 import java.util.List;