+++ swing/common-ui/src/main/java/module-info.java
@@ -31,6 +31,7 @@
  * <li>{@link is.codion.swing.common.ui.dialog.Dialogs}
  * <li>{@link is.codion.swing.common.ui.key.KeyEvents}
  * <li>{@link is.codion.swing.common.ui.laf.LookAndFeelProvider}
+ * <li>{@link is.codion.swing.common.ui.laf.LookAndFeelProviders}
  * <li>{@link is.codion.swing.common.ui.layout.Layouts}
  * </ul>
  */
@@ -66,4 +67,6 @@ module is.codion.swing.common.ui {
 	exports is.codion.swing.common.ui.key;
 	exports is.codion.swing.common.ui.laf;
 	exports is.codion.swing.common.ui.layout;
+
+	uses is.codion.swing.common.ui.laf.LookAndFeelProviders;
 }