+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityApplicationPanel.java
@@ -53,6 +53,7 @@ import is.codion.swing.common.ui.control.ToggleControl;
 import is.codion.swing.common.ui.dialog.Dialogs;
 import is.codion.swing.common.ui.laf.LookAndFeelComboBox;
 import is.codion.swing.common.ui.laf.LookAndFeelProvider;
+import is.codion.swing.common.ui.laf.LookAndFeelProviders;
 import is.codion.swing.common.ui.layout.Layouts;
 import is.codion.swing.framework.model.SwingEntityApplicationModel;
 import is.codion.swing.framework.ui.EntityPanel.WindowType;
@@ -656,8 +657,8 @@ public abstract class EntityApplicationPanel<M extends SwingEntityApplicationMod
 	/**
 	 * Allows the user the select between the available Look and Feels, saves the selection as a user preference.
 	 * @return a Control for selecting the application look and feel
-	 * @see LookAndFeelProvider#addLookAndFeel(LookAndFeelProvider)
-	 * @see LookAndFeelProvider#findLookAndFeelProvider(String)
+	 * @see LookAndFeelProviders#addLookAndFeel(LookAndFeelProvider)
+	 * @see LookAndFeelProviders#findLookAndFeelProvider(String)
 	 * @see Dialogs#lookAndFeelSelectionDialog()
 	 * @see LookAndFeelComboBox#ENABLE_ON_SELECTION
 	 */
@@ -1286,7 +1287,7 @@ public abstract class EntityApplicationPanel<M extends SwingEntityApplicationMod
 		/**
 		 * Sets the default look and feel classname, used in case no look and feel settings are found in user preferences.
 		 * Note that for an external Look and Feels to be enabled, it must be registered via
-		 * {@link LookAndFeelProvider#addLookAndFeel(LookAndFeelProvider)}
+		 * {@link LookAndFeelProviders#addLookAndFeel(LookAndFeelProvider)}
 		 * before starting the application.
 		 * @param defaultLookAndFeelClassName the default look and feel classname
 		 * @return this Builder instance
@@ -1296,7 +1297,7 @@ public abstract class EntityApplicationPanel<M extends SwingEntityApplicationMod
 		/**
 		 * Sets the look and feel classname, overrides any look and feel settings found in user preferences.
 		 * Note that for an external Look and Feels to be enabled, it must be registered via
-		 * {@link LookAndFeelProvider#addLookAndFeel(LookAndFeelProvider)}
+		 * {@link LookAndFeelProviders#addLookAndFeel(LookAndFeelProvider)}
 		 * before starting the application.
 		 * @param lookAndFeelClassName the look and feel classname
 		 * @return this Builder instance