+++ changelog.md
@@ -10,10 +10,12 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - LookAndFeelPanel.initializeLookAndFeelDefaults() error message improved.
 - ExceptionDialogBuilder.WRAPPER_EXCEPTIONS, RuntimeException re-added, but unwrapping now unwraps according to the exact class, not isAssignableFrom, RemoteException removed.
+- LookAndFeelProviders added, ServiceLoader based, related refactoring.
 ### is.codion.swing.framework.ui 
 - DefaultEntityApplicationPanelBuilder no longer validates that the look and feel exists on the classpath, should not prevent application start.
 ### is.codion.plugin.intellij.themes
 - Added a bunch of IntelliJ theme based look and feels along with all of those available in com.formdev.flatlaf-intellij-themes.
+- IntelliJThemes now implements LookAndFeelProviders.
 
 ## 0.18.23
 ### is.codion.common